feat : 경매 등록 api 적용 성공
